+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Organize the data sources and classify the pills by region on the left navigation of the explore page.
   [GEOD-1086](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1086)
+- Mobile Detail Page Improvements
+  [IMLCZO-230](https://opensource.ncsa.illinois.edu/jira/browse/IMLCZO-230)
 
 ### Fixed
 - When clicking on a pill in explore left navigation, open the popup on the map